// src/shoppinghistoryservice/src/generated/shoppinghistoryservice.ts
export interface CreateOrderRequest {
  user_id: string;
  positions: OrderItem[];
}

export interface OrderItem {
  product_id: string;
  quantity: number;
}

export interface GetOrderRequest {
  order_id: string;
}

export interface GetOrdersOfUserRequest {
  user_id: string;
}

export interface DeleteOrderRequest {
  order_id: string;
}

export interface Order {
  order_id: string;
  user_id: string;
  created_at: Date | undefined;
  order_items: OrderItem[];
}

export interface Orders {
  orders: Order[];
}

export interface Empty {
}

function createBaseCreateOrderRequest(): CreateOrderRequest {
  return { user_id: "", positions: [] };
}

export const CreateOrderRequest: MessageFns<CreateOrderRequest> = {
  encode(message: CreateOrderRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.user_id !== "") {
      writer.uint32(10).string(message.user_id);
    }
    for (const v of message.positions) {
      OrderItem.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateOrderRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateOrderRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.user_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.positions.push(OrderItem.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateOrderRequest {
    return {
      user_id: isSet(object.user_id) ? globalThis.String(object.user_id) : "",
      positions: globalThis.Array.isArray(object?.positions)
        ? object.positions.map((e: any) => OrderItem.fromJSON(e))
        : [],
    };
  },

  toJSON(message: CreateOrderRequest): unknown {
    const obj: any = {};
    if (message.user_id !== "") {
      obj.user_id = message.user_id;
    }
    if (message.positions?.length) {
      obj.positions = message.positions.map((e) => OrderItem.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateOrderRequest>, I>>(base?: I): CreateOrderRequest {
    return CreateOrderRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateOrderRequest>, I>>(object: I): CreateOrderRequest {
    const message = createBaseCreateOrderRequest();
    message.user_id = object.user_id ?? "";
    message.positions = object.positions?.map((e) => OrderItem.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const GetOrderRequest: MessageFns<GetOrderRequest> = {
  encode(message: GetOrderRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.order_id !== "") {
      writer.uint32(10).string(message.order_id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetOrderRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetOrderRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.order_id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetOrderRequest {
    return { order_id: isSet(object.order_id) ? globalThis.String(object.order_id) : "" };
  },

  toJSON(message: GetOrderRequest): unknown {
    const obj: any = {};
    if (message.order_id !== "") {
      obj.order_id = message.order_id;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetOrderRequest>, I>>(base?: I): GetOrderRequest {
    return GetOrderRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetOrderRequest>, I>>(object: I): GetOrderRequest {
    const message = createBaseGetOrderRequest();
    message.order_id = object.order_id ?? "";
    return message;
  },
};

function createBaseGetOrdersOfUserRequest(): GetOrdersOfUserRequest {
  return { user_id: "" };
}

export const GetOrdersOfUserRequest: MessageFns<GetOrdersOfUserRequest> = {
  encode(message: GetOrdersOfUserRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.user_id !== "") {
      writer.uint32(10).string(message.user_id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetOrdersOfUserRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetOrdersOfUserRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.user_id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetOrdersOfUserRequest {
    return { user_id: isSet(object.user_id) ? globalThis.String(object.user_id) : "" };
  },

  toJSON(message: GetOrdersOfUserRequest): unknown {
    const obj: any = {};
    if (message.user_id !== "") {
      obj.user_id = message.user_id;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetOrdersOfUserRequest>, I>>(base?: I): GetOrdersOfUserRequest {
    return GetOrdersOfUserRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetOrdersOfUserRequest>, I>>(object: I): GetOrdersOfUserRequest {
    const message = createBaseGetOrdersOfUserRequest();
    message.user_id = object.user_id ?? "";
    return message;
  },
};

function createBaseDeleteOrderRequest(): DeleteOrderRequest {
  return { order_id: "" };
}

export const DeleteOrderRequest: MessageFns<DeleteOrderRequest> = {
  encode(message: DeleteOrderRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.order_id !== "") {
      writer.uint32(10).string(message.order_id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteOrderRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteOrderRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.order_id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DeleteOrderRequest {
    return { order_id: isSet(object.order_id) ? globalThis.String(object.order_id) : "" };
  },

  toJSON(message: DeleteOrderRequest): unknown {
    const obj: any = {};
    if (message.order_id !== "") {
      obj.order_id = message.order_id;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DeleteOrderRequest>, I>>(base?: I): DeleteOrderRequest {
    return DeleteOrderRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DeleteOrderRequest>, I>>(object: I): DeleteOrderRequest {
    const message = createBaseDeleteOrderRequest();
    message.order_id = object.order_id ?? "";
    return message;
  },
};

function createBaseOrder(): Order {
  return { order_id: "", user_id: "", created_at: undefined, order_items: [] };
}

export const Order: MessageFns<Order> = {
  encode(message: Order, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.order_id !== "") {
      writer.uint32(10).string(message.order_id);
    }
    if (message.user_id !== "") {
      writer.uint32(18).string(message.user_id);
    }
    if (message.created_at !== undefined) {
      Timestamp.encode(toTimestamp(message.created_at), writer.uint32(26).fork()).join();
    }
    for (const v of message.order_items) {
      OrderItem.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Order {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOrder();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.order_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.user_id = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.created_at = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.order_items.push(OrderItem.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Order {
    return {
      order_id: isSet(object.order_id) ? globalThis.String(object.order_id) : "",
      user_id: isSet(object.user_id) ? globalThis.String(object.user_id) : "",
      created_at: isSet(object.created_at) ? fromJsonTimestamp(object.created_at) : undefined,
      order_items: globalThis.Array.isArray(object?.order_items)
        ? object.order_items.map((e: any) => OrderItem.fromJSON(e))
        : [],
    };
  },

  toJSON(message: Order): unknown {
    const obj: any = {};
    if (message.order_id !== "") {
      obj.order_id = message.order_id;
    }
    if (message.user_id !== "") {
      obj.user_id = message.user_id;
    }
    if (message.created_at !== undefined) {
      obj.created_at = message.created_at.toISOString();
    }
    if (message.order_items?.length) {
      obj.order_items = message.order_items.map((e) => OrderItem.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Order>, I>>(base?: I): Order {
    return Order.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Order>, I>>(object: I): Order {
    const message = createBaseOrder();
    message.order_id = object.order_id ?? "";
    message.user_id = object.user_id ?? "";
    message.created_at = object.created_at ?? undefined;
    message.order_items = object.order_items?.map((e) => OrderItem.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const Empty: MessageFns<Empty> = {
  encode(_: Empty, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Empty {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEmpty();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): Empty {
    return {};
  },

  toJSON(_: Empty): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<Empty>, I>>(base?: I): Empty {
    return Empty.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Empty>, I>>(_: I): Empty {
    const message = createBaseEmpty();
    return message;
  },
};